feat(patches): add refresh button in view and command on the cmd palette
Also brings:
- refactor existing hacky implementation of patches' TreeDataProvider
etc into it's own files
