+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🚀 Enhancements
 
 - **patches**: show new view in the sidebar listing all Radicle Patches, including a tooltip with more info on hover ([#43](https://github.com/cytechmobile/radicle-vscode-extension/issues/43))
+- **commands**: new VS Code command to refresh the list of Radicle Patches ([#43](https://github.com/cytechmobile/radicle-vscode-extension/issues/43))
 - **ux:** convert existing flows with series of user input steps into a bespoke, cohesive experience indicating total step count and completion progress, allow editing of already submitted answers of the same flow, as well reveal-toggling of typed-in passwords, among other improvements ([#67](https://github.com/cytechmobile/radicle-vscode-extension/issues/67))
 - **auth:** additionally show the alias associated with a Radicle identity whenever showing the latter ([#67](https://github.com/cytechmobile/radicle-vscode-extension/issues/67))
