+++ src/helpers/index.ts
@@ -5,3 +5,4 @@ export * from './fileWatcher'
 export * from './httpd'
 export * from './logExtensionActivated'
 export * from './radCli'
+export * from './views'
