+++ src/ux/index.ts
@@ -1,4 +1,5 @@
 export * from './httpdConnection'
+export * from './patches'
 export * from './radCliInstallation'
 export * from './radicleIdentityAuth'
 export * from './radicleProject'
